+++ lib/reform/contract.rb
@@ -91,22 +91,6 @@ module Reform
     require "reform/schema"
     extend Reform::Schema
   end
-
-  class Contract_ # DISCUSS: make class?
-    # allows including representers from Representable, Roar or disposable.
-    def self.inherit_module!(representer) # called from Representable::included.
-      # representer_class.inherit_module!(representer)
-      representer.representable_attrs.each do |dfn|
-        next if dfn.name == "links" # wait a second # FIXME what is that?
-
-        # TODO: remove manifesting and do that in representable, too!
-        args = [dfn.name, dfn.instance_variable_get(:@options)] # TODO: dfn.to_args (inluding &block)
-
-        property(*args) and next unless dfn.representer_module
-        property(*args) { include dfn.representer_module } # nested.
-      end
-    end
-  end
 end
 
 require 'reform/contract/errors'