+++ lib/reform/form/model_reflections.rb
@@ -30,7 +30,7 @@ module Reform::Form::ModelReflections
   def defined_enums
     return model.defined_enums unless is_a?(Reform::Form::Composition)
 
-    model.each.with_object({}) { |m,h| h.merge! m.defined_enums }
+    mapper.each.with_object({}) { |m,h| h.merge! m.defined_enums }
   end
 
   # this should also contain to_param and friends as this is used by the form helpers.